- Compile warning fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33828 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/gui/cegui/ceguigeometrybuffer.cpp
@@ -135,7 +135,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(cegui)
 
       simple.object2world = d_matrix.GetTransform();
 
-      simple.indexStart = indexBuf.GetSize();
+      simple.indexStart = (int)indexBuf.GetSize();
       simple.indexEnd = simple.indexStart;
 
       renderMeshes.Push(simple);
@@ -150,7 +150,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(cegui)
     {
       const CEGUI::Vertex& vs = vbuff[i];
       // convert from CEGUI::Vertex to something directly usable by CS.
-      indexBuf.Push(vertBuf.GetSize());
+      indexBuf.Push((int)vertBuf.GetSize());
       csVector3 v(vs.position.d_x + d_texelOffset.d_x, 
                   vs.position.d_y + d_texelOffset.d_y, 
                   vs.position.d_z);
@@ -195,13 +195,13 @@ CS_PLUGIN_NAMESPACE_BEGIN(cegui)
   //----------------------------------------------------------------------------//
   uint GeometryBuffer::getVertexCount() const
   {
-    return vertBuf.GetSize();
+    return (int)vertBuf.GetSize();
   }
 
   //----------------------------------------------------------------------------//
   uint GeometryBuffer::getBatchCount() const
   {
-    return renderMeshes.GetSize();
+    return (int)renderMeshes.GetSize();
   }
 
   //----------------------------------------------------------------------------//
